+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## [UNRELEASED]
+
+- support for headers in streamable MCP servers
+
 ## [0.23.0]
 
 - fixed laggy prompt field clearing after submitting message